+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ### Release 4.3.4
 * Optimize activity picker, so users can see whether or not the used link comes from an activity
+* Carousel caption now can handle images
 * Fix hyphenation for wrapping long words - Thanks to Florian Dagner (#37)
 * Fix missing draftitemid in carousel